Modify maxcompute data type for float (#638)
# Description
The MR modifies datatype of prediction log table from float to Double
and enable print logs by adding env variable

# Modifications
1. modifies datatype of prediction log table from float to Double

The incoming prediction logs are stored in a temp table and the temp
table is inserted to the prediction log table
The temp table is created using df.persist() which only supports double
and bigint (and not float and int)
So it is necessary for the prediction log table to be using double type
to maintain the compatibility and avoid casting

2. Adds Environment variable to print logs

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

+++ python/observation-publisher/publisher/observation_sink.py
@@ -441,7 +441,7 @@ class MaxComputeSink(ObservationSink):
     def schema_fields(self) -> List[Column]:
         value_type_to_maxcompute_type = {
             ValueType.INT64: "BIGINT",
-            ValueType.FLOAT64: "FLOAT",
+            ValueType.FLOAT64: "DOUBLE",
             ValueType.BOOLEAN: "BOOLEAN",
             ValueType.STRING: "STRING",
         }
@@ -542,7 +542,6 @@ class MaxComputeSink(ObservationSink):
                         f"Errors when inserting rows to MaxCompute, retrying attempt {i}/{self.retry.retry_attempts}"
                     )
                     time.sleep(self.retry.retry_interval_seconds)
-            
         print(f"Failed to write to MaxCompute after {self.retry.retry_attempts} attempts")
